fix: remaining ESLint issues blocking deploy
- Move getTimeRange utility out of TimeRangeSelector component file
  to fix react-refresh/only-export-components violation
- Fix DependencyDetailPage setState-in-effect
