+++ web/src/pages/DependenciesPage.tsx
@@ -1,7 +1,8 @@
 import { useState, useEffect, useCallback, useMemo, type ReactElement } from 'react'
 import { api } from '../api/client'
 import type { DependencySummary } from '../api/types'
-import { TimeRangeSelector, getTimeRange } from '../components/TimeRangeSelector'
+import { TimeRangeSelector } from '../components/TimeRangeSelector'
+import { getTimeRange } from '../utils/timeRange'
 import { AutoRefresh } from '../components/AutoRefresh'
 import { formatDuration, formatErrorRate, errorRateColor, formatCount } from '../utils/format'
 import { DependencyDetailPanel } from './DependencyDetailPage'
